chore: migrated helper.py tests to dedicated module
